+++ CHANGES.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- The character selection menu now only shows characters which are included in the
+  player's slot.
+- Locked characters now use their character portrait instead of the "padlock" icon to
+  make it easier to see which characters are available.
+  - The portraits are still faded to indicate they are not selectable yet.
+- The locked panel now displays the text "Missing AP item for <character name>" instead
+  of the base game requirement to unlock the character.
 - When connecting to an Archipelago server, various information is now logged to help
   with debugging.
   - The information is logged to the Brotato mod log file at